get rid of billing v1 stuff
