+++ frontend/src/scenes/billing/BillingGauge.scss
