--- frontend/src/scenes/billing/billingLogic.ts
@@ -1,236 +1,0 @@
-import { kea, path, actions, connect, reducers, selectors, events, listeners } from 'kea'
-import { loaders } from 'kea-loaders'
-import api from 'lib/api'
-import type { billingLogicType } from './billingLogicType'
-import { PlanInterface, BillingType } from '~/types'
-import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
-import posthog from 'posthog-js'
-import { sceneLogic } from 'scenes/sceneLogic'
-import { Scene } from 'scenes/sceneTypes'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { windowValues } from 'kea-window-values'
-import { getBreakpoint } from 'lib/utils/responsiveUtils'
-import { urlToAction } from 'kea-router'
-import { urls } from 'scenes/urls'
-import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
-import { billingV2Logic } from './v2/billingV2Logic'
-
-export const UTM_TAGS = 'utm_medium=in-product&utm_campaign=billing-management'
-export const ALLOCATION_THRESHOLD_ALERT = 0.85 // Threshold to show warning of event usage near limit
-
-export enum BillingAlertType {
-    SetupBilling = 'setup_billing',
-    UsageNearLimit = 'usage_near_limit',
-    UsageLimitExceeded = 'usage_limit_exceeded',
-}
-
-export const billingLogic = kea<billingLogicType>([
-    path(['scenes', 'billing', 'billingLogic']),
-    actions({
-        registerInstrumentationProps: true,
-        toggleUsageTiers: true,
-        setPlans: (plans: PlanInterface[]) => ({ plans }),
-        referer: (referer: string) => ({ referer }),
-    }),
-    connect({
-        values: [preflightLogic, ['preflight'], featureFlagLogic, ['featureFlags'], billingV2Logic, ['billingVersion']],
-        actions: [eventUsageLogic, ['reportIngestionBillingCancelled']],
-    }),
-    reducers({
-        showUsageTiers: [
-            false as boolean,
-            {
-                toggleUsageTiers: (state) => !state,
-            },
-        ],
-        billingSuccessRedirect: [
-            urls.projectHomepage() as string,
-            {
-                referer: (_, { referer }) => {
-                    if (referer === 'ingestion') {
-                        return urls.events()
-                    } else {
-                        return urls.projectHomepage()
-                    }
-                },
-            },
-        ],
-    }),
-    windowValues({
-        isSmallScreen: (window: Window) => window.innerWidth < getBreakpoint('md'),
-    }),
-    loaders(({ actions }) => ({
-        billing: [
-            null as BillingType | null,
-            {
-                loadBilling: async () => {
-                    const response = await api.get('api/billing/')
-                    if (!response?.plan) {
-                        actions.loadPlans()
-                    } else {
-                        actions.setPlans([response.plan])
-                    }
-                    actions.registerInstrumentationProps()
-                    return response as BillingType
-                },
-                setBillingLimit: async (billing: BillingType) => {
-                    const res = await api.update('api/billing/', billing)
-                    lemonToast.success(`Billing limit set to $${billing.billing_limit} usd/month`)
-
-                    return res as BillingType
-                },
-            },
-        ],
-        plans: [
-            [] as PlanInterface[],
-            {
-                loadPlans: async () => {
-                    const response = await api.get('api/plans?self_serve=1')
-                    return response.results
-                },
-                setPlans: ({ plans }) => plans,
-            },
-        ],
-        billingSubscription: [
-            null as BillingType | null,
-            {
-                subscribe: async (plan) => {
-                    return await api.create('billing/subscribe', { plan })
-                },
-            },
-        ],
-        planDetails: [
-            null as string | null,
-            {
-                loadPlanDetails: async (plan) => {
-                    const response = await fetch(`/api/plans/${plan}/template/`)
-                    if (response.ok) {
-                        return await response.text()
-                    }
-                    return null
-                },
-            },
-        ],
-    })),
-    selectors({
-        upgradeLink: [
-            (s) => [s.preflight, s.billingVersion],
-            (preflight, billingVersion): string =>
-                billingVersion === 'v2' || preflight?.cloud
-                    ? '/organization/billing'
-                    : 'https://license.posthog.com?utm_medium=in-product&utm_campaign=in-product-upgrade',
-        ],
-        eventAllocation: [(s) => [s.billing], (billing: BillingType) => billing?.event_allocation],
-        percentage: [
-            (s) => [s.eventAllocation, s.billing],
-            (eventAllocation: number | null, billing: BillingType) => {
-                if (!eventAllocation || !billing?.current_usage) {
-                    return null
-                }
-                return Math.min(Math.round((billing.current_usage / eventAllocation) * 100) / 100, 1)
-            },
-        ],
-        strokeColor: [
-            (selectors) => [selectors.percentage],
-            (percentage) => {
-                let color: string | Record<string, string> = 'var(--primary)'
-                if (percentage === null || percentage === undefined) {
-                    /* No event limit set */
-                    color = {
-                        from: '#1890FF',
-                        to: '#52C41A',
-                    }
-                }
-
-                if (percentage && percentage > 0.65 && percentage < 0.8) {
-                    color = 'var(--warning)'
-                }
-                if (percentage && percentage > 0.8) {
-                    color = 'var(--danger)'
-                }
-                return color
-            },
-        ],
-        alertToShow: [
-            (s) => [s.eventAllocation, s.percentage, s.billing, sceneLogic.selectors.scene, s.billingVersion],
-            (
-                eventAllocation: number | null,
-                percentage: number,
-                billing: BillingType,
-                scene: Scene,
-                billingVersion: string
-            ): BillingAlertType | undefined => {
-                if (billingVersion === 'v2') {
-                    return
-                }
-                if (billing?.should_setup_billing && billing?.subscription_url) {
-                    // Determines which billing alert/warning to show to the user (if any)
-
-                    // Priority 1: In-progress incomplete billing setup
-                    return BillingAlertType.SetupBilling
-                }
-
-                // Priority 2: Event allowance exceeded or near limit
-                if (billing?.billing_limit_exceeded) {
-                    return BillingAlertType.UsageLimitExceeded
-                }
-
-                // Priority 3: Event allowance near threshold
-                if (
-                    scene !== Scene.Billing &&
-                    billing?.is_billing_active &&
-                    billing?.current_usage &&
-                    eventAllocation &&
-                    percentage >= ALLOCATION_THRESHOLD_ALERT
-                ) {
-                    return BillingAlertType.UsageNearLimit
-                }
-            },
-        ],
-    }),
-    events(({ actions }) => ({
-        afterMount: () => {
-            if (preflightLogic.values.preflight?.cloud) {
-                actions.loadBilling()
-            }
-        },
-    })),
-    listeners(({ values }) => ({
-        subscribeSuccess: ({ billingSubscription }) => {
-            if (billingSubscription?.subscription_url) {
-                window.location.href = billingSubscription.subscription_url
-            }
-        },
-        registerInstrumentationProps: async (_, breakpoint) => {
-            await breakpoint(100)
-            if (posthog && values.billing) {
-                posthog.register({
-                    has_billing_plan: !!values.billing?.plan,
-                    metered_billing: values.billing.plan?.is_metered_billing,
-                    event_allocation: values.billing.event_allocation,
-                    allocation_used:
-                        values.billing.event_allocation && values.billing.current_usage !== null
-                            ? values.billing.current_usage / values.billing.event_allocation
-                            : undefined,
-                })
-            }
-        },
-
-        loadBillingSuccess: () => {
-            if (!values.billing) {
-                return
-            }
-        },
-    })),
-    urlToAction(({ actions }) => ({
-        '/ingestion/billing': (_, params) => {
-            if (params.reason === 'cancelled') {
-                actions.reportIngestionBillingCancelled()
-            }
-        },
-        '/organization/billing/subscribed': (_, { referer }) => {
-            actions.referer(referer)
-        },
-    })),
-])
