+++ frontend/src/scenes/billing/billingV2Logic.ts
@@ -8,7 +8,7 @@ import { urlToAction } from 'kea-router'
 import { forms } from 'kea-forms'
 import { dayjs } from 'lib/dayjs'
 import { lemonToast } from '@posthog/lemon-ui'
-import { projectUsage } from './billing-utils'
+import { projectUsage } from './v2/billing-utils'
 import posthog from 'posthog-js'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { userLogic } from 'scenes/userLogic'
